Updated failing execution tests: semanticTests/{revertStrings, array/copying}.

+++ test/libsolidity/semanticTests/revertStrings/calldata_array_invalid_length.sol
@@ -1,6 +1,6 @@
 pragma abicoder               v2;
 contract C {
-	function f(uint256[][] calldata x) external returns (uint256) {
+	function f(uint256[][] calldata x) external returns (uint) {
 		return x[0].length;
 	}
 }
@@ -9,4 +9,4 @@ contract C {
 // revertStrings: debug
 // compileViaYul: also
 // ----
-// f(uint256[][]): 0x20, 1, 0x20, 0x0100000000000000000000 -> FAILURE, hex"08c379a0", 0x20, 28, "Invalid calldata tail length"
+// f(uint256[][]): refargs { 0x01, 0x01, 0x01, 0x10, 0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f } -> 0x10
